lesson-1.3/helphub.ai/challenges/challenge-1/requirements.txt
Flask==2.3.3
openai>=1.6.0
python-dotenv==1.0.0